templated AssetBundle
