+++ src/AssetBundle.cpp
@@ -1,7 +1,4 @@
-#include "utils.hpp"
-#include "utils-logging.hpp"
-#include "beatsaber-hook/shared/utils/il2cpp-functions.hpp"
-#include "beatsaber-hook/shared/utils/il2cpp-utils.hpp"
+#include "AssetBundle.hpp"
 
 namespace bs_utils {
     bool CheckAssetClass(Il2CppObject* asset) {
@@ -16,16 +13,6 @@ namespace bs_utils {
         return true;
     }
 
-    void AssetBundle::AssetComplete(AssetCallback callback, Il2CppObject* assetBundleRequest) {
-        getLogger().info("AssetBundle::AssetComplete called!");
-        CRASH_UNLESS(il2cpp_utils::GetPropertyValue<bool>(assetBundleRequest, "isDone").value_or(false));
-        auto* asset = RET_V_UNLESS(il2cpp_utils::GetPropertyValue<Asset*>(assetBundleRequest, "asset"));
-        // RET_V_UNLESS(CheckAssetClass(asset));
-        callback(asset);
-        il2cpp_utils::RunMethod(assetBundleRequest, "Finalize");
-        getLogger().info("AssetBundle::AssetComplete finished!");
-    }
-
     Asset* AssetBundle::LoadAsset(std::string_view assetName, Il2CppReflectionType* assetType) {
         if (!assetType) assetType = CRASH_UNLESS(il2cpp_utils::GetSystemType("UnityEngine", "GameObject"));
         auto* nameStr = CRASH_UNLESS(il2cpp_utils::createcsstr(assetName));
@@ -34,30 +21,7 @@ namespace bs_utils {
         return RET_0_UNLESS(il2cpp_utils::RunMethod<Asset*>(this, "LoadAsset", nameStr, assetType));
     }
 
-    bool AssetBundle::LoadAssetAsync(std::string_view assetName, AssetCallback callback, Il2CppReflectionType* assetType) {
-        if (!assetType) assetType = CRASH_UNLESS(il2cpp_utils::GetSystemType("UnityEngine", "GameObject"));
-        auto* nameStr = CRASH_UNLESS(il2cpp_utils::createcsstr(assetName));
-        getLogger().info("Loading asset '%s' asynchronously.", assetName.data());
-
-        auto* assetAsync = RET_0_UNLESS(il2cpp_utils::RunMethod(this, "LoadAssetAsync", nameStr, assetType));
-
-        auto* method = RET_0_UNLESS(il2cpp_utils::FindMethodUnsafe(assetAsync, "add_completed", 1));
-        auto* action = RET_0_UNLESS(il2cpp_utils::MakeAction(method, 0, (void*)callback, AssetComplete));
-
-        RET_0_UNLESS(il2cpp_utils::RunMethod(assetAsync, method, action));
-        getLogger().info("Began loading asset async");
-        return true;
-    }
-
-    void AssetBundle::AssetBundleComplete(AssetBundleCallback callback, Il2CppObject* assetBundleCreateRequest) {
-        getLogger().info("UnityAssetLoader: AssetBundleCreateRequestComplete Called!");
-        auto* assetBundle = RET_V_UNLESS(il2cpp_utils::GetPropertyValue<AssetBundle*>(assetBundleCreateRequest, "assetBundle"));
-        callback(assetBundle);
-        il2cpp_utils::RunMethod(assetBundleCreateRequest, "Finalize");
-        getLogger().info("UnityAssetLoader: AssetBundleCreateRequestComplete Finished!");
-    }
-
-    Il2CppString* CreateIl2CppStringPathAndEnsureExists(std::string_view filePath) {
+    Il2CppString* AssetBundle::CreateIl2CppStringPathAndEnsureExists(std::string_view filePath) {
         auto* assetFilePath = RET_0_UNLESS(il2cpp_utils::createcsstr(filePath));
         bool pathExists = RET_0_UNLESS(il2cpp_utils::RunMethod<bool>("System.IO", "File", "Exists", assetFilePath));
         RET_0_UNLESS(pathExists);
@@ -67,19 +31,5 @@ namespace bs_utils {
     AssetBundle* AssetBundle::LoadFromFile(std::string_view filePath) {
         auto* assetFilePath = RET_0_UNLESS(CreateIl2CppStringPathAndEnsureExists(filePath));
         return RET_0_UNLESS(il2cpp_utils::RunMethod<AssetBundle*>("UnityEngine", "AssetBundle", "LoadFromFile", assetFilePath));
-    }
-
-    bool AssetBundle::LoadFromFileAsync(std::string_view filePath, AssetBundleCallback callback) {
-        auto* assetFilePath = RET_0_UNLESS(CreateIl2CppStringPathAndEnsureExists(filePath));
-
-        auto* bundleAsync = RET_0_UNLESS(il2cpp_utils::RunMethod("UnityEngine", "AssetBundle", "LoadFromFileAsync", assetFilePath));
-        RET_0_UNLESS(il2cpp_utils::SetPropertyValue(bundleAsync, "allowSceneActivation", true));
-
-        auto method = RET_0_UNLESS(il2cpp_utils::FindMethodUnsafe(bundleAsync, "add_completed", 1));
-        auto action = RET_0_UNLESS(il2cpp_utils::MakeAction(method, 0, (void*)callback, AssetBundleComplete));
-
-        RET_0_UNLESS(il2cpp_utils::RunMethod(bundleAsync, method, action));
-        getLogger().info("Began loading asset bundle '%s' async.", filePath.data());
-        return true;
     }
 }
